backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.1
pydantic==2.9.2
pydantic-settings==2.5.2
bcrypt==4.2.0
email-validator==2.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
greenlet==3.0.3
